requirements.txt
python-dotenv==1.0.0
pandas==2.1.0